Se agregaro el valor saldo a la base de datos

+++ app/src/main/java/com/example/suletitoapp/Usuario.kt
@@ -5,5 +5,5 @@ data class Usuario(
     var apellidos: String = "",
     var rol: String = "",
     var telefono: String = "",
-    var saldo: Int = 0
+    var saldo: Double = 0.0
 )
